fe: add secondary font
There are 2 things included in this commit:

- Secondary font is added.
- The tailwind config is refactored so that the fonts declaring is
abstracted out of the config object to make it easy on the eye. However,
somehow the fontFamily constant cannot be in other file and imported
into the tailwind config file without causing the typescript's
intellisense function to not working correctly. Therefore, the font
constant object is staying the config file for now.